+++ src/pages/map/index.tsx
@@ -206,6 +206,7 @@ const Map: NextPage = () => {
         <title>탐색 | Slam - 우리 주변 농구장을 빠르게</title>
       </Head>
       <DatePicker
+        startDate={today}
         selectedDate={selectedDateAndSlot.selectedDate}
         onClick={handleDateClick}
       />
